Fix bug in institutional units
Institutional units was displaying admins on it's show page for all it's
departments as well as it's unit admins. This is because we were calling
the parent class administrator to get the unit's admins which also
contained all of it's child department admins. We only wanted the unit
admins so we updated the controller/model/view accordingly.